% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proposals_votes.R
\name{get_proposals_votes}
\alias{get_proposals_votes}
\title{Scrape votes for tagging proposals}
\usage{
get_proposals_votes(urls)
}
\arguments{
\item{urls}{a string containing the tagging proposal to retrieve voting history from.}
}
\value{
a dataframe with the following columns:
\itemize{
\item \code{fullurl}: (string) the URL to the tagging proposal wiki page.
\item \code{votes_raw}: (string) the raw text describing the vote.
\item \code{vote}: (factor) vote (Approve, Oppose, Abstain, Other), based on the svg starting the \code{title} attribute of the vote icon's link.
\item \code{user}: (factor) voter's username.
\item \code{date_vote}: (POSIXct, UTC) date and time the vote was cast.
}
}
\description{
Given a vector of urls, scrapes the tagging proposals to retrieve all the votes received.
}
\examples{
proposal_votes <- get_proposals_votes('https://wiki.openstreetmap.org/wiki/Proposal:Electricity')

proposal_votes

}
