% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proposals}
\alias{proposals}
\title{OSM Tagging Proposals Data}
\format{
columns:
\describe{
\item{status}{(factor) the current status of the proposal: Approved, Rejected, Voting, Proposed, Draft, Abandoned, Canceled, Obsoleted, Inactive.}
\item{title}{(string) the title of the page containing the proposal.}
\item{sortkeyprefix}{a string}
\item{timestamp}{a date and time.}
\item{pagelanguage}{(factor) the page's language, according to WikiMedia. Note that this will not necessarily match the actual language used to write the proposal. This is because how MediaWiki handles internationalisation, which assumes that every page is first written in English and then translated into other languages. Because that, any page without translation will be considered to be in English.}
\item{touched}{a timestamp when the page was last touched}
\item{length}{(integer) page's length, in bytes.}
\item{fullurl}{(string) URL pointing to the proposal's page.}
\item{editurl}{(string) the URL to edit the page.}
\item{pageid}{(int) the page's ID.}
\item{page_creator}{(factor) the username who created the page.}
\item{date_of_page_creation}{(date) the date and time in which the page was created.}
\item{latest_editor}{(factor) the username who last edited the page.}
\item{date_of_latest_edit}{(date) the date and time in which the page was last updated.}
\item{total_number_of_edits}{(integer) number of total times that the page has been edited.}
\item{total_number_of_disctinct_authors}{(integer) number of distinct users who have edited the page.}
\item{proposal_status}{(factor) status of a proposal.}
\item{proposed_by}{(string) name of the user(s) making the proposal. This can contain multiple values and therefore, differ from page_creator. This is because there can only be one user creating a page, but the proposals can be a joint effort.}
\item{tagging}{(string) short description of the proposed tagging scheme.}
\item{applies_to_node}{(Boolean) whether this tag applies to nodes or not.}
\item{applies_to_way}{(Boolean) whether this tag applies to ways or not.}
\item{applies_to_area}{(Boolean) whether this tag applies to areas or not.}
\item{applies_to_relation}{(Boolean) whether this tag applies to relations or not.}
\item{definition}{(string) short definition of what the tagging proposal aims to describe.}
\item{rendered_as}{a string describing how the tagg is/should be rendered in the map (if any).}
\item{draft_started}{date in which the proposal draft started.}
\item{rfc_start}{date in which Request For Comments (RFC) started.}
\item{vote_start}{date in which the voting process started.}
\item{vote_end}{date in which the voting process ended.}

}
}
\source{
\url{https://wiki.openstreetmap.org/wiki/Proposal_process}
}
\usage{
proposals
}
\description{
A data frame with 2115 rows and 30 columns, containing all
Approved, Rejected and Proposed tagging proposals from the \href{https://wiki.openstreetmap.org}{OSM Wiki}, as well as their associated metadata.
This dataset has been created by the function \code{get_tagging_proposals()}
ran from the script \code{data-raw/tagging_proposals.R}.
}
\details{
This documentation is not yet complete: lacks all the final fields from the dataset.
}
\keyword{datasets}
