% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proposals_votes}
\alias{proposals_votes}
\title{OSM Tagging Proposals' votes Data}
\format{
columns: \describe{ \item{url}{(factor, string) the URL to the tagging
proposal wiki page.} \item{votes_raw}{a string with the wiki page containing
the proposal.} \item{vote}{(factor) vote type (approve, abstain, oppose).}
\item{user}{(factor) voter's username} \item{date_vote}{(POSIXct, UTC) date
and time in which the vote was cast.}

}
}
\source{
\url{https://wiki.openstreetmap.org/wiki/Proposal_process#Voting}
}
\usage{
proposals_votes
}
\description{
A data frame with 8511 rows and
5 columns, containing all the votes cast for every
tagging proposal in \code{proposals} marked as "Approved", "Rejected", "Proposed"
and "Cancelled" (currently consisting of 812
different proposals). This dataset is generated by the function
\code{get_proposals_votes()}, ran from the script \code{data-raw/tagging_proposals.R}.
}
\keyword{datasets}
